Refactor search functionality
