+++ src/pages/search.tsx
@@ -1,37 +1,33 @@
-import { ChangeEvent, useCallback, useEffect, useMemo, useState } from 'react';
+import { useState } from 'react';
 import { NextSeo } from 'next-seo';
 import type { NextPage, GetStaticProps } from 'next';
-import debounce from 'lodash.debounce';
 import NextImage from 'next/image';
 import {
     GridLayout,
     TextInput,
     ESystemIconNames,
     Button,
-    Link,
     TypographyScale,
 } from '@mdb/flora';
-import useSWR from 'swr';
 
 import Hero from '../components/hero';
 import { DesktopFilters, MobileFilters } from '../components/search-filters';
 import { pageWrapper } from '../styled/styles';
 
 import { FilterItem } from '../components/search-filters';
-import { fetcher } from '../components/search/utils';
-import { getFilters } from '../page-templates/content-type/utils';
+import { getFilters } from '../hooks/search/utils';
 import {
     resultsStringAndTagsStyles,
     desktopFiltersStyles,
 } from '../page-templates/content-type/styles';
 import { searchBoxStyles } from '../components/search/styles';
-import FilterTag from '../page-templates/content-type/filter-tag';
 
 import Results from '../components/search/results';
 import { useRouter } from 'next/router';
-import { itemInFilters } from '../page-templates/content-type/utils';
 import { getURLPath } from '../utils/format-url-path';
 import { thumbnailLoader } from '../components/card/utils';
+import useSearch from '../hooks/search';
+import FilterTagSection from '../components/search-filters/filter-tag-section';
 
 export interface SearchProps {
     l1Items: FilterItem[];
@@ -51,267 +47,43 @@ const Search: NextPage<SearchProps> = ({
     expertiseLevelItems,
 }) => {
     const router = useRouter();
-    const { isReady, query } = router;
 
-    const [allFilters, setAllFilters] = useState<FilterItem[]>([]);
-    const [searchString, setSearchString] = useState<string>('');
-    const [resultstoShow, setResultsToShow] = useState<number>(10);
-    const { data, error, isValidating } = useSWR(
-        () => `s=${searchString}`,
-        fetcher,
-        {
-            revalidateIfStale: false,
-            revalidateOnFocus: false,
-            revalidateOnReconnect: false,
-            shouldRetryOnError: false,
-        }
-    );
+    const {
+        data,
+        error,
+        isValidating,
+        fullyLoaded,
+        setResultsToShow,
+        resultsToShow,
+        allFilters,
+        setAllFilters,
+        onSearch,
+        onFilter,
+        searchString,
+        setSearchString,
+        numberOfResults,
+    } = useSearch(undefined, undefined, router, {
+        l1Items,
+        languageItems,
+        technologyItems,
+        contributedByItems,
+        contentTypeItems,
+        expertiseLevelItems,
+    });
+
     const [filterTagsExpanded, setFilterTagsExpanded] = useState(false);
     const [mobileFiltersOpen, setMobileFiltersOpen] = useState(false);
-
-    // Populate the search/filters with query params on page load/param change.
-    useEffect(() => {
-        if (isReady) {
-            const {
-                s,
-                product,
-                language,
-                technology,
-                contributedBy,
-                contentType,
-                expertiseLevel,
-            } = query;
-            if (s && typeof s === 'string' && s !== searchString) {
-                setSearchString(s);
-            }
-            let allNewFilters: FilterItem[] = [];
-            if (product) {
-                // Gotta look for L1s and L2s that match.
-                const products =
-                    typeof product === 'object' ? product : [product];
-                let productFilters: FilterItem[] = [];
-
-                products.forEach(prod => {
-                    const filterProduct = l1Items.find(
-                        l1 =>
-                            l1.name === prod ||
-                            l1.subItems.find(l2 => l2.name === prod)
-                    );
-                    if (filterProduct) {
-                        if (filterProduct.name !== prod) {
-                            // This means it's an L2 match.
-                            productFilters.push(
-                                filterProduct.subItems.find(
-                                    l2 => l2.name === prod
-                                ) as FilterItem
-                            );
-                        } else {
-                            productFilters.push(filterProduct);
-                        }
-                    }
-                });
-                allNewFilters = allNewFilters.concat(productFilters);
-            }
-            if (contentType) {
-                // Gotta look for content types and dig down into the subcategories of Code Examples.
-                const contentTypes =
-                    typeof contentType === 'object'
-                        ? contentType
-                        : [contentType];
-                let contentTypeFilters: FilterItem[] = [];
-
-                contentTypes.forEach(ct => {
-                    const filterContentType = contentTypeItems.find(
-                        l1 =>
-                            l1.name === ct ||
-                            l1.subItems.find(l2 => l2.name === ct)
-                    );
-                    if (filterContentType) {
-                        if (filterContentType.name !== ct) {
-                            // This means it's an L2 match.
-                            contentTypeFilters.push(
-                                filterContentType.subItems.find(
-                                    l2 => l2.name === ct
-                                ) as FilterItem
-                            );
-                        } else {
-                            contentTypeFilters.push(filterContentType);
-                        }
-                    }
-                });
-                allNewFilters = allNewFilters.concat(contentTypeFilters);
-            }
-            // For the rest, just map it to the corresponding item.
-            if (language) {
-                // Technically can either come in as a string of a string[], so convert to a string[]
-                // and loop over by default.
-                const languages =
-                    typeof language === 'object' ? language : [language];
-                const languageFilters = languageItems.filter(lang =>
-                    languages.includes(lang.name)
-                );
-                allNewFilters = allNewFilters.concat(languageFilters);
-            }
-            if (technology) {
-                const technologies =
-                    typeof technology === 'object' ? technology : [technology];
-                const technologyFilters = technologyItems.filter(tech =>
-                    technologies.includes(tech.name)
-                );
-                allNewFilters = allNewFilters.concat(technologyFilters);
-            }
-            if (contributedBy) {
-                const contributedBys =
-                    typeof contributedBy === 'object'
-                        ? contributedBy
-                        : [contributedBy];
-                const contributedByFilters = contributedByItems.filter(
-                    contrib => contributedBys.includes(contrib.name)
-                );
-                allNewFilters = allNewFilters.concat(contributedByFilters);
-            }
-            if (expertiseLevel) {
-                const expertiseLevels =
-                    typeof expertiseLevel === 'object'
-                        ? expertiseLevel
-                        : [expertiseLevel];
-                const expertiseLevelFilters = expertiseLevelItems.filter(exp =>
-                    expertiseLevels.includes(exp.name)
-                );
-                allNewFilters = allNewFilters.concat(expertiseLevelFilters);
-            }
-            setAllFilters(allNewFilters);
-        }
-    }, [isReady]); // Missing query dependency, but that's ok because we only need this on first page load.
-
-    const onFilter = useCallback((filters: FilterItem[]) => {
-        setResultsToShow(10);
-        setAllFilters(filters);
-        const product = filters
-            .filter(
-                filter =>
-                    filter.type === 'L1Product' || filter.type === 'L2Product'
-            )
-            .map(filter => filter.name);
-        const language = filters
-            .filter(filter => filter.type === 'ProgrammingLanguage')
-            .map(filter => filter.name);
-        const technology = filters
-            .filter(filter => filter.type === 'Technology')
-            .map(filter => filter.name);
-        const contentType = filters
-            .filter(
-                filter =>
-                    filter.type === 'ContentType' || filter.type === 'CodeLevel'
-            )
-            .map(filter => filter.name);
-        const contributedBy = filters
-            .filter(filter => filter.type === 'AuthorType')
-            .map(filter => filter.name);
-        const expertiseLevel = filters
-            .filter(filter => filter.type === 'ExpertiseLevel')
-            .map(filter => filter.name);
-
-        const query = {
-            s: searchString,
-            product,
-            language,
-            technology,
-            contentType,
-            contributedBy,
-            expertiseLevel,
-        };
-        router.replace({ pathname: '/search', query }, undefined, {
-            scroll: false,
-            shallow: true,
-        });
-    }, []);
 
     const clearFilters = () => {
         onFilter([]);
     };
 
-    const onSearch = (event: ChangeEvent<HTMLInputElement>) => {
-        setResultsToShow(10);
-        setSearchString(event.target.value);
-        // Have to preserve the filters here as well.
-        const product = allFilters
-            .filter(
-                filter =>
-                    filter.type === 'L1Product' || filter.type === 'L2Product'
-            )
-            .map(filter => filter.name);
-        const language = allFilters
-            .filter(filter => filter.type === 'ProgrammingLanguage')
-            .map(filter => filter.name);
-        const technology = allFilters
-            .filter(filter => filter.type === 'Technology')
-            .map(filter => filter.name);
-        const contentType = allFilters
-            .filter(
-                filter =>
-                    filter.type === 'ContentType' || filter.type === 'CodeLevel'
-            )
-            .map(filter => filter.name);
-        const contributedBy = allFilters
-            .filter(filter => filter.type === 'AuthorType')
-            .map(filter => filter.name);
-        const expertiseLevel = allFilters
-            .filter(filter => filter.type === 'ExpertiseLevel')
-            .map(filter => filter.name);
-
-        router.replace(
-            {
-                pathname: '/search',
-                query: {
-                    s: event.target.value,
-                    product,
-                    language,
-                    technology,
-                    contentType,
-                    contributedBy,
-                    expertiseLevel,
-                },
-            },
-            undefined,
-            { scroll: false, shallow: true }
-        );
-    };
-
-    const debouncedOnSearch = useMemo(
-        () => debounce(onSearch, 400), // Not sure what this value should be.
-        [allFilters]
-    );
-    // Stop the invocation of the debounced function
-    // after unmounting
-    useEffect(() => {
-        return () => {
-            debouncedOnSearch.cancel();
-        };
-    }, []);
-
-    const onFilterTabClose = (filterTag: FilterItem) => {
+    const onFilterTagClose = (filterTag: FilterItem) => {
         const newFilters = allFilters.filter(filter => filter !== filterTag);
         onFilter(newFilters);
     };
 
     const hasFiltersSet = !!allFilters.length;
-
-    const filteredData = (() => {
-        if (!data) {
-            return [];
-        } else if (!hasFiltersSet) {
-            return data;
-        } else {
-            return data.filter(item => {
-                return itemInFilters(item, allFilters);
-            });
-        }
-    })();
-
-    const numberOfResults = filteredData.length;
-    const shownData = filteredData.slice(0, resultstoShow);
-    const fullyLoaded = resultstoShow >= numberOfResults;
 
     ///////////////////////////////////////
     // COMPUTED ELEMENTS
@@ -352,68 +124,6 @@ const Search: NextPage<SearchProps> = ({
         </div>
     );
 
-    const filterTags = (
-        <div
-            sx={{
-                gap: 'inc30',
-                flexWrap: 'wrap',
-                display: ['none', null, null, 'flex'],
-            }}
-        >
-            {allFilters.length > 5 ? (
-                filterTagsExpanded ? (
-                    <>
-                        {allFilters.map(filter => (
-                            <FilterTag
-                                key={`${filter.name} ${filter.type}`}
-                                filter={filter}
-                                closeIcon={true}
-                                onClick={onFilterTabClose}
-                            />
-                        ))}
-                        <FilterTag
-                            filter={{
-                                name: 'Show less',
-                                count: 0,
-                                subItems: [],
-                            }}
-                            onClick={() => setFilterTagsExpanded(false)}
-                        />
-                    </>
-                ) : (
-                    <>
-                        {allFilters.slice(0, 5).map(filter => (
-                            <FilterTag
-                                key={`${filter.name} ${filter.type}`}
-                                filter={filter}
-                                closeIcon={true}
-                                onClick={onFilterTabClose}
-                            />
-                        ))}
-                        <FilterTag
-                            filter={{
-                                name: `+${allFilters.length - 5}`,
-                                count: 0,
-                                subItems: [],
-                            }}
-                            onClick={() => setFilterTagsExpanded(true)}
-                        />
-                    </>
-                )
-            ) : (
-                allFilters.map(filter => (
-                    <FilterTag
-                        key={`${filter.name} ${filter.type}`}
-                        filter={filter}
-                        closeIcon={true}
-                        onClick={onFilterTabClose}
-                    />
-                ))
-            )}
-            <Link onClick={clearFilters}>Clear all filters</Link>
-        </div>
-    );
-
     const resultsStringAndTags = (
         <div sx={resultsStringAndTagsStyles}>
             {(data || isValidating) && (
@@ -430,7 +140,15 @@ const Search: NextPage<SearchProps> = ({
                         : ''}
                 </TypographyScale>
             )}
-            {hasFiltersSet && filterTags}
+            {hasFiltersSet && (
+                <FilterTagSection
+                    allFilters={allFilters}
+                    filterTagsExpanded={filterTagsExpanded}
+                    setFilterTagsExpanded={setFilterTagsExpanded}
+                    onFilterTagClose={onFilterTagClose}
+                    clearFilters={clearFilters}
+                />
+            )}
             <div
                 sx={{
                     display: ['block', null, null, 'none'],
@@ -491,15 +209,15 @@ const Search: NextPage<SearchProps> = ({
                                 label="Search All"
                                 iconName={ESystemIconNames.SEARCH}
                                 value={searchString}
-                                onChange={debouncedOnSearch}
+                                onChange={onSearch}
                                 autoFocus={true}
                             />
                         </div>
                         {resultsStringAndTags}
-                        {!!filteredData.length || isValidating || error ? (
+                        {!!data.length || isValidating || error ? (
                             <>
                                 <Results
-                                    data={shownData}
+                                    data={data}
                                     isLoading={isValidating}
                                     hasError={error}
                                 />
@@ -515,7 +233,7 @@ const Search: NextPage<SearchProps> = ({
                                             <Button
                                                 onClick={() =>
                                                     setResultsToShow(
-                                                        resultstoShow + 10
+                                                        resultsToShow + 10
                                                     )
                                                 }
                                                 variant="secondary"
